# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Scheme {
  id: string;
  name: string;
  purpose: string;
  maxLoan: number;
  interest: number;
  moratorium: number;
  tenure: number;
  applicantType: string;
  tags: string[];
  isPrototypeData: boolean;
}

export type ApplicantProfileApplicantType = typeof ApplicantProfileApplicantType[keyof typeof ApplicantProfileApplicantType];


export const ApplicantProfileApplicantType = {
  Entrepreneur: 'Entrepreneur',
  Student: 'Student',
} as const;

export interface ApplicantProfile {
  name?: string;
  applicantType: ApplicantProfileApplicantType;
  /** @minimum 0 */
  age: number;
  category: string;
  education?: string;
  employment?: string;
  /** @minimum 0 */
  income: number;
  purpose: string;
  /** @minimum 0 */
  projectCost?: number;
  /** @minimum 0 */
  loanRequired: number;
  location: string;
}

export type SchemeMatchBreakdown = {[key: string]: number};

export type SchemeMatch = Scheme & {
  score: number;
  reasons: string[];
  breakdown: SchemeMatchBreakdown;
};

export interface EmiInput {
  /** @minimum 0 */
  principal: number;
  /** @minimum 0 */
  annualRate: number;
  /** @minimum 0 */
  tenureYears: number;
  /** @minimum 0 */
  moratoriumMonths: number;
}

export interface EmiResult {
  emi: number;
  principal: number;
  totalInterest: number;
  totalRepayment: number;
}

export interface Partner {
  id: string;
  name: string;
  type: string;
  distance: number;
  supportedSchemes: string[];
  serviceArea: string;
  status: string;
  processingCapacity: number;
  fundingAvailability: number;
  address: string;
  latitude?: number;
  longitude?: number;
}

export interface PartnerRequest {
  schemeId: string;
  location: string;
}

export type PartnerMatch = Partner & {
  score: number;
  reasons: string[];
};

export interface ApplicationInput {
  name: string;
  phone: string;
  email: string;
  address: string;
  schemeId: string;
  partnerId: string;
  income: number;
  loanRequired: number;
  projectType: string;
  projectCost: number;
  documents?: string[];
}

export type Application = ApplicationInput & {
  id: string;
  status: string;
  createdAt: string;
};

export type AdminAnalyticsMonthlyItem = {
  month: string;
  applications: number;
};

export type AdminAnalyticsDemandItem = {
  name: string;
  value: number;
};

export type AdminAnalyticsStatusItem = {
  name: string;
  value: number;
};

export type AdminAnalyticsPartnersItem = {
  name: string;
  applications: number;
};

export interface AdminAnalytics {
  totalApplications: number;
  matchedApplications: number;
  routedApplications: number;
  assistance: number;
  monthly: AdminAnalyticsMonthlyItem[];
  demand: AdminAnalyticsDemandItem[];
  status: AdminAnalyticsStatusItem[];
  partners: AdminAnalyticsPartnersItem[];
}

/**
 * Not found
 */
export type NotFoundResponse = Error;