# lib/api-zod/src/generated/types/adminAnalytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */
import type { AdminAnalyticsDemandItem } from './adminAnalyticsDemandItem';
import type { AdminAnalyticsMonthlyItem } from './adminAnalyticsMonthlyItem';
import type { AdminAnalyticsPartnersItem } from './adminAnalyticsPartnersItem';
import type { AdminAnalyticsStatusItem } from './adminAnalyticsStatusItem';

export interface AdminAnalytics {
  totalApplications: number;
  matchedApplications: number;
  routedApplications: number;
  assistance: number;
  monthly: AdminAnalyticsMonthlyItem[];
  demand: AdminAnalyticsDemandItem[];
  status: AdminAnalyticsStatusItem[];
  partners: AdminAnalyticsPartnersItem[];
}