# lib/api-zod/src/generated/types/applicantProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */
import type { ApplicantProfileApplicantType } from './applicantProfileApplicantType';

export interface ApplicantProfile {
  name?: string;
  applicantType: ApplicantProfileApplicantType;
  /** @minimum 0 */
  age: number;
  category: string;
  education?: string;
  employment?: string;
  /** @minimum 0 */
  income: number;
  purpose: string;
  /** @minimum 0 */
  projectCost?: number;
  /** @minimum 0 */
  loanRequired: number;
  location: string;
}