# lib/api-zod/src/generated/types/applicationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */

export interface ApplicationInput {
  name: string;
  phone: string;
  email: string;
  address: string;
  schemeId: string;
  partnerId: string;
  income: number;
  loanRequired: number;
  projectType: string;
  projectCost: number;
  documents?: string[];
}