# lib/api-zod/src/generated/types/emiInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */

export interface EmiInput {
  /** @minimum 0 */
  principal: number;
  /** @minimum 0 */
  annualRate: number;
  /** @minimum 0 */
  tenureYears: number;
  /** @minimum 0 */
  moratoriumMonths: number;
}