# lib/api-zod/src/generated/types/emiResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */

export interface EmiResult {
  emi: number;
  principal: number;
  totalInterest: number;
  totalRepayment: number;
}