# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */

export interface Partner {
  id: string;
  name: string;
  type: string;
  distance: number;
  supportedSchemes: string[];
  serviceArea: string;
  status: string;
  processingCapacity: number;
  fundingAvailability: number;
  address: string;
  latitude?: number;
  longitude?: number;
}