# lib/api-zod/src/generated/types/partnerMatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */
import type { Partner } from './partner';

export type PartnerMatch = Partner & {
  score: number;
  reasons: string[];
};