# lib/api-zod/src/generated/types/scheme.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */

export interface Scheme {
  id: string;
  name: string;
  purpose: string;
  maxLoan: number;
  interest: number;
  moratorium: number;
  tenure: number;
  applicantType: string;
  tags: string[];
  isPrototypeData: boolean;
}