# lib/api-zod/src/generated/types/schemeMatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SchemeSathi prototype API for scheme matching and application routing
 * OpenAPI spec version: 0.1.0
 */
import type { Scheme } from './scheme';
import type { SchemeMatchBreakdown } from './schemeMatchBreakdown';

export type SchemeMatch = Scheme & {
  score: number;
  reasons: string[];
  breakdown: SchemeMatchBreakdown;
};